% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{coef.ebnm}
\alias{coef.ebnm}
\title{Extract posterior means from a fitted EBNM model}
\usage{
\method{coef}{ebnm}(object, ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\description{
The \code{\link[stats]{coef}} method for class \code{\link{ebnm}}.
}
