% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sampler.R
\name{ebnm_add_sampler}
\alias{ebnm_add_sampler}
\title{Add sampler to an ebnm_object}
\usage{
ebnm_add_sampler(ebnm_res)
}
\arguments{
\item{ebnm_res}{The fitted \code{ebnm} object.}
}
\value{
The \code{ebnm} object with an additional field
  \code{posterior_sampler}.
}
\description{
Adds a posterior sampler to a fitted \code{\link{ebnm}} object.
}
