% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_fns.R
\name{ebnm_ash}
\alias{ebnm_ash}
\title{Solve the EBNM problem using an ash family of distributions}
\usage{
ebnm_ash(
  x,
  s = 1,
  mode = 0,
  scale = "estimate",
  g_init = NULL,
  fix_g = FALSE,
  output = output_default(),
  control = NULL,
  ...
)
}
\arguments{
\item{x}{A vector of observations. Missing observations (\code{NA}s) are
not allowed.}

\item{s}{A vector of standard errors (or a scalar if all are equal).
Standard errors may not be exactly zero, and missing standard errors are
not allowed.}

\item{mode}{Passed to \code{\link[ashr]{ash}} as parameter \code{mode}.}

\item{scale}{Passed to \code{\link[ashr]{ash}} as parameter \code{mixsd}.}

\item{g_init}{Passed to \code{\link[ashr]{ash}} as parameter \code{g}.}

\item{fix_g}{Passed to \code{\link[ashr]{ash}} as parameter \code{fixg}.}

\item{output}{A character vector indicating which values are to be returned.
Function \code{output_default()} provides the default return values, while
\code{output_all()} lists all possible return values. See \strong{Value}
below.}

\item{control}{A list of control parameters to be passed to optimization
function \code{\link[mixsqp]{mixsqp}}.}

\item{...}{Additional parameters to be passed to \code{\link[ashr]{ash}}.}
}
\value{
An \code{ebnm} object. Depending on the argument to \code{output}, the
  object is a list containing elements:
    \describe{
      \item{\code{data}}{A data frame containing the observations \code{x}
        and standard errors \code{s}.}
      \item{\code{posterior}}{A data frame of summary results (posterior
        means, standard deviations, second moments, and local false sign
        rates).}
      \item{\code{fitted_g}}{The fitted prior \eqn{\hat{g}}.}
      \item{\code{log_likelihood}}{The optimal log likelihood attained,
        \eqn{L(\hat{g})}.}
      \item{\code{posterior_sampler}}{A function that can be used to
        produce samples from the posterior. The sampler takes a single
        parameter \code{nsamp}, the number of posterior samples to return per
        observation.}
     }
   S3 methods \code{confint}, \code{fitted}, \code{logLik}, \code{nobs},
   \code{plot}, \code{predict}, \code{print}, \code{samp}, and \code{summary}
   have been implemented for \code{ebnm} objects. For details, see the
   respective help pages, linked below under \strong{See Also}.
}
\description{
A wrapper to function \code{\link[ashr]{ash}} in package \code{ashr}.
  Identical to function \code{\link{ebnm}} with argument
  \code{prior_family = "ash"}.
}
\seealso{
Available S3 methods include \code{\link{confint.ebnm}},
   \code{\link{fitted.ebnm}}, \code{\link{logLik.ebnm}},
   \code{\link{nobs.ebnm}}, \code{\link{plot.ebnm}},
   \code{\link{predict.ebnm}},
   \code{\link{print.ebnm}}, \code{\link{print.summary.ebnm}},
   \code{\link{samp.ebnm}}, and \code{\link{summary.ebnm}}.
}
