% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_fns.R
\name{ebnm_deconvolver}
\alias{ebnm_deconvolver}
\title{Solve the EBNM problem using the "deconvolveR" family of distributions}
\usage{
ebnm_deconvolver(
  x,
  s = 1,
  scale = "estimate",
  g_init = NULL,
  fix_g = FALSE,
  output = output_default(),
  control = NULL,
  ...
)
}
\arguments{
\item{x}{A vector of observations. Missing observations (\code{NA}s) are
not allowed.}

\item{s}{Standard errors, which must be uniformly equal to 1 (i.e.,
\code{s = 1}) since the deconvolveR method takes \eqn{z}-scores as input.}

\item{scale}{A deconvolveR prior is a finite mixture of point masses
\deqn{\pi_1 \delta_{\mu_1} + \ldots + \pi_K \delta_{\mu_K},}
where parameters \eqn{\pi_k} are estimated and the point masses are
evenly spaced over \eqn{(\mu_1, \mu_K)}.The distance between successive
point masses can be specified by the user via parameter
\code{scale}, in which case the argument should be a scalar specifying the
distance \eqn{d = \mu_2 - \mu_1 = \cdots = \mu_K - \mu_{K - 1}};
alternatively, if \code{scale = "estimate"}, then \code{ebnm} sets the grid
via function \code{\link{init_g_for_npmle}}.}

\item{g_init}{The prior distribution \eqn{g}. Usually this is left
unspecified (\code{NULL}) and estimated from the data. However, it can be
used in conjuction with \code{fix_g = TRUE} to fix the prior (useful, for
example, to do computations with the "true" \eqn{g} in simulations). If
\code{g_init} is specified but \code{fix_g = FALSE}, \code{g_init}
specifies the initial value of \eqn{g} used during optimization. This has
the side effect of fixing the \code{scale} parameter. When supplied,
\code{g_init} should be an object of class \code{\link[ashr]{normalmix}}.}

\item{fix_g}{If \code{TRUE}, fix the prior \eqn{g} at \code{g_init} instead
of estimating it.}

\item{output}{A character vector indicating which values are to be returned.
Function \code{output_default()} provides the default return values, while
\code{output_all()} lists all possible return values. See \strong{Value}
below.}

\item{control}{A list of control parameters to be passed to optimization
function \code{\link[stats]{nlm}}.}

\item{...}{Additional parameters to be passed to function
\code{\link[deconvolveR]{deconv}} in package \code{deconvolveR}.}
}
\value{
An \code{ebnm} object. Depending on the argument to \code{output}, the
  object is a list containing elements:
    \describe{
      \item{\code{data}}{A data frame containing the observations \code{x}
        and standard errors \code{s}.}
      \item{\code{posterior}}{A data frame of summary results (posterior
        means, standard deviations, second moments, and local false sign
        rates).}
      \item{\code{fitted_g}}{The fitted prior \eqn{\hat{g}}.}
      \item{\code{log_likelihood}}{The optimal log likelihood attained,
        \eqn{L(\hat{g})}.}
      \item{\code{posterior_sampler}}{A function that can be used to
        produce samples from the posterior. The sampler takes a single
        parameter \code{nsamp}, the number of posterior samples to return per
        observation.}
     }
   S3 methods \code{confint}, \code{fitted}, \code{logLik}, \code{nobs},
   \code{plot}, \code{predict}, \code{print}, \code{samp}, and \code{summary}
   have been implemented for \code{ebnm} objects. For details, see the
   respective help pages, linked below under \strong{See Also}.
}
\description{
Solves the empirical Bayes normal means (EBNM) problem using a non-parametric
  exponential family with a natural spline basis.
  Like \code{\link{ebnm_npmle}}, there is no unimodal assumption, but whereas
  \code{ebnm_npmle} produces spiky estimates for \eqn{g},
  \code{ebnm_deconvolver} estimates are much more regular. See
  \code{\link[deconvolveR]{deconvolveR-package}} for details and
  references. Identical to function \code{\link{ebnm}} with argument
  \code{prior_family = "deconvolver"}.
}
\seealso{
Available S3 methods include \code{\link{confint.ebnm}},
   \code{\link{fitted.ebnm}}, \code{\link{logLik.ebnm}},
   \code{\link{nobs.ebnm}}, \code{\link{plot.ebnm}},
   \code{\link{predict.ebnm}},
   \code{\link{print.ebnm}}, \code{\link{print.summary.ebnm}},
   \code{\link{samp.ebnm}}, and \code{\link{summary.ebnm}}.
}
