% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm.R
\name{ebnm_point_laplace}
\alias{ebnm_point_laplace}
\title{Solve the Empirical Bayes Normal Means problem with point-laplace prior}
\usage{
ebnm_point_laplace(x, s = 1)
}
\arguments{
\item{x}{a vector of observations}

\item{s}{a vector of standard deviations (or scalar if all equal)}
}
\value{
a list with elements result, fitted_g, and loglik
}
\details{
Given vectors of data x, and standard errors s, solve EBNM problem with "point-laplace" prior
(i.e. a mixture of point mass at 0 and Laplace distribution).
That is the prior is pi0 \delta_0 + (1-pi0)DExp(a) where Dexp is the double exponential
(Laplace) distribution, and (pi0,a) are esimated by marginal maximum likelihood. The model and code are based on
EbayesThresh by Johnstone and Silverman, but simplified by removing thresholding, and
improved numerical stability.
}
\examples{
mu = c(rep(0,1000), rexp(1000)) # means
s = rgamma(2000,1,1) #standard errors
x = mu + rnorm(2000,0,s) # observations
x.ebnm = ebnm_point_laplace(x,s)
ashr::get_pm(x.ebnm) # posterior mean

}
