% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eight_schools}
\alias{eight_schools}
\title{Eight Schools Data}
\format{
A data frame with 8 rows and 2 columns:
\describe{
  \item{x}{The treatment effect estimate.}
  \item{s}{The standard error.}
}
}
\source{
Rubin DB (1981). "Estimation in Parallel Randomized Experiments."
  \emph{Journal of Educational Statistics}, \strong{6}(4), 377-401.
}
\usage{
eight_schools
}
\description{
The eight schools dataset analyzed by Rubin (1981) and Gelman et al. (2014).
In each of eight US high schools, different randomized experiments were
conducted in order to estimate the effect of coaching programs on
standardized test scores. The data includes estimates of treatment effects
as well as standard errors.
}
\seealso{
Gelman A, Carlin JB, Stern HS, Dunson DB, Vehtari A, Rubin DB (2014).
\emph{Bayesian data analysis}. Texts in Statistical Science Series, third
edition. CRC Press, Boca Raton, FL.
}
\keyword{datasets}
