% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_point_laplace.R
\name{genlaplacemix}
\alias{genlaplacemix}
\title{Constructor for genlaplacemix class}
\usage{
genlaplacemix(pi, mean, scale_pos, scale_neg)
}
\arguments{
\item{pi}{A vector of mixture proportions.}

\item{mean}{A vector of means.}

\item{scale_pos}{A vector of positive scale parameters.}

\item{scale_neg}{A vector of negative scale parameters.}
}
\value{
An object of class \code{genlaplacemix} (a list with elements
  \code{pi}, \code{mean}, \code{scale_pos}, and \code{scale_neg}, described above).
}
\description{
Creates a finite mixture of Generalized Laplace distributions.
}
