% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/horseshoe.R
\name{horseshoe}
\alias{horseshoe}
\title{Constructor for horseshoe class}
\usage{
horseshoe(scale)
}
\arguments{
\item{scale}{The scale parameter (must be a scalar).}
}
\value{
An object of class \code{horseshoe} (a list with a single element
  \code{scale}, described above).
}
\description{
Creates a horseshoe prior (see Carvalho, Polson, and Scott (2010)). The
  horseshoe is usually parametrized as
  \eqn{\theta_i \sim N(0, s^2 \tau^2 \lambda_i^2)},
  \eqn{\lambda_i \sim \mathrm{Cauchy}^+(0, 1)},
  with \eqn{s^2} the variance of the error distribution. We use a single
  parameter \code{scale}, which corresponds to \eqn{s\tau} and thus does
  not depend on the error distribution.
}
