% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{logLik.ebnm}
\alias{logLik.ebnm}
\title{Extract the log likelihood from a fitted EBNM model}
\usage{
\method{logLik}{ebnm}(object, ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\value{
An object of class \code{logLik}, which includes attributes \code{df},
  the degrees of freedom --- i.e., number of parameters in the model ---, and
  \code{nobs}, the number of observations in the data.
}
\description{
The \code{\link[stats]{logLik}} method for class \code{\link{ebnm}}.
}
