% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_laplace.R
\name{mle_laplace}
\alias{mle_laplace}
\title{Compute mle of w and a for data (x,s) under point laplace prior}
\usage{
mle_laplace(x, s, control = NULL)
}
\details{
Does simple mle (currently using optim)
Compared with the original wandafromx this function dispenses with the
complication of the thresholds
}
