% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_normal.R
\name{mle_normal}
\alias{mle_normal}
\title{Compute mle of w and a for data (x,s) under point normal prior}
\usage{
mle_normal(x, s, startpar = NULL, control = NULL)
}
\arguments{
\item{x}{observations}

\item{s}{standard deviations}

\item{startpar}{initialization}

\item{control}{list of parameters to be passed to optim}
}
\description{
Paragraph-length description goes here.
}
\details{
Does simple mle (currently using optim)
}
