% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{nobs.ebnm}
\alias{nobs.ebnm}
\title{Get the number of observations used to fit an EBNM model}
\usage{
\method{nobs}{ebnm}(object, ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\value{
The number of observations used to fit the \code{ebnm} object.
}
\description{
The \code{\link[stats]{nobs}} method for class \code{\link{ebnm}}.
}
