% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{plot.ebnm}
\alias{plot.ebnm}
\title{Plot an ebnm object}
\usage{
\method{plot}{ebnm}(x, ..., remove_abline = FALSE)
}
\arguments{
\item{x}{The fitted \code{ebnm} object.}

\item{...}{Additional \code{ebnm} objects to be included on the same plot.}

\item{remove_abline}{To better illustrate shrinkage effects, the plot
includes the line \eqn{y = x} by default. If
\code{remove_abline = TRUE}, then this line will not be drawn.}
}
\value{
A \code{ggplot} object.
}
\description{
Given one or more fitted \code{ebnm} object(s), produces a plot of posterior
  means vs. observations.
}
\details{
An object of class \code{ggplot} is returned, so that the plot can be
  customized in the usual \code{\link[ggplot2]{ggplot2}} fashion.
}
\examples{
theta <- c(rep(0, 100), rexp(100))
theta[1:50] <- 0
s <- 1
x <- theta + rnorm(200, 0, s)
pn.res <- ebnm_point_normal(x, s)
plot(pn.res)

pe.res <- ebnm_point_exponential(x, s)
plot(pn.res, pe.res)

# Customize plot:
library(ggplot2)
plot(pn.res, pe.res, remove_abline = TRUE) +
  theme_bw() +
  labs(x = "Simulated data")

}
