% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_sampler_normal.R
\name{post_sampler_normal}
\alias{post_sampler_normal}
\title{Sample from the posterior for data (x,s) under point normal prior}
\usage{
post_sampler_normal(x, s, w, a, nsamp)
}
\arguments{
\item{x}{observations}

\item{s}{standard deviations}

\item{w}{weight of non-null component in fitted prior}

\item{a}{variance of non-null component in fitted prior}

\item{nsamp}{number of samples to return per observation}
}
