% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{predict.ebnm}
\alias{predict.ebnm}
\title{Use the estimated prior from a fitted EBNM model to solve the EBNM problem for
  new data}
\usage{
\method{predict}{ebnm}(object, newdata, output = output_default(), ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{newdata}{A list that includes fields \code{x} (the new observations)
and \code{s} (corresponding standard errors). See \code{\link{ebnm}} for
details about how to specify \code{x} and \code{s}.}

\item{output}{A character vector indicating which values are to be returned.
See \code{\link{ebnm}} for details.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\value{
An \code{ebnm} object. See \code{\link{ebnm}} for details.
}
\description{
The \code{\link[stats]{predict}} method for class \code{\link{ebnm}}.
}
