% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{print.summary.ebnm}
\alias{print.summary.ebnm}
\title{Print a summary.ebnm object}
\usage{
\method{print}{summary.ebnm}(x, digits = 2, ...)
}
\arguments{
\item{x}{The \code{summary.ebnm} object.}

\item{digits}{Number of significant digits to use.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\description{
The \code{print} method for class \code{\link{summary.ebnm}}.
}
