% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{summary.ebnm}
\alias{summary.ebnm}
\title{Summarize an ebnm object}
\usage{
\method{summary}{ebnm}(object, ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\value{
A \code{summary.ebnm} object.
}
\description{
The \code{summary} method for class \code{\link{ebnm}}.
}
