% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_bm.R
\name{tree_bm}
\alias{tree_bm}
\title{Constructor for tree_bm prior class}
\usage{
tree_bm(bm_var)
}
\arguments{
\item{bm_var}{A positive scalar: the Brownian motion variance per unit
branch length.}
}
\value{
An object of class \code{tree_bm}.
}
\description{
Creates a Brownian motion tree prior object storing the estimated diffusion
variance.
}
