% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_ou.R
\name{tree_ou}
\alias{tree_ou}
\title{Constructor for tree_ou prior class}
\usage{
tree_ou(alpha, stationary_var)
}
\arguments{
\item{alpha}{A positive scalar giving the OU reversion rate per unit
branch length.}

\item{stationary_var}{A positive scalar giving the stationary variance of
the OU process.}
}
\value{
An object of class \code{tree_ou}.
}
\description{
Creates an Ornstein-Uhlenbeck tree prior object storing the estimated
reversion rate and stationary variance.
}
