% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generalized_binary.R
\name{truncnormmix}
\alias{truncnormmix}
\title{Constructor for truncnormmix class}
\usage{
truncnormmix(pi, a, b, mean, sd)
}
\arguments{
\item{pi}{A vector of mixture proportions.}

\item{a}{A vector of lower bounds. These may be \code{-Inf}.}

\item{b}{A vector of upper bounds. These may be \code{Inf}.}

\item{mean}{A vector of (untruncated) means.}

\item{sd}{A vector of (untruncated) standard deviations.}
}
\value{
An object of class \code{truncnormmix} (a list with elements
  \code{pi}, \code{a}, \code{b}, \code{mean}, and \code{sd}, described above).
}
\description{
Creates a finite mixture of truncated normal distributions.
}
