% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{vcov.ebnm}
\alias{vcov.ebnm}
\title{Extract posterior variances from a fitted EBNM model}
\usage{
\method{vcov}{ebnm}(object, ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\description{
The \code{\link[stats]{vcov}} method for class \code{\link{ebnm}}.
}
