% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wandafromx.mle.R
\name{wandafromx.mle}
\alias{wandafromx.mle}
\title{Compute mle of w and a for data (x,s)}
\usage{
wandafromx.mle(x, s)
}
\details{
Does simple mle (currently using optim)
Compared with the original wandafromx this function dispenses with the
complication of the thresholds
}
