% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm.R
\name{ebnm}
\alias{ebnm}
\title{Solve the Empirical Bayes Normal Means problem}
\usage{
ebnm(x, s = 1)
}
\arguments{
\item{x}{a vector of observations}

\item{s}{a vector of standard deviations (or scalar if all equal)}
}
\value{
a list with elements result, fitted_g, and loglik
}
\details{
Given vectors of data x, and standard errors s, fit the point laplace prior
(i.e. a mixture of point mass at 0 and Laplace distribution)
by fitting the two parameters (pi0 and laplace rate). The model and code are based on
EbayesThresh by Johnstone and Silverman, but simplified by removing thresholding, and
improved numerical stability
}
