% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_point_laplace.R
\name{ebnm_point_laplace}
\alias{ebnm_point_laplace}
\title{Solve the Empirical Bayes Normal Means problem with point-laplace prior}
\usage{
ebnm_point_laplace(x, s = 1, g = NULL, fixg = FALSE, output = NULL)
}
\arguments{
\item{x}{a vector of observations}

\item{s}{a vector of standard deviations (or scalar if all equal)}

\item{g}{The prior distribution (list with elements pi0,a). Usually this is unspecified (NULL) and
estimated from the data. However, it can be used in conjuction with fixg=TRUE
to specify the g to use (e.g. useful in simulations to do computations with the "true" g).
Or, if g is specified but fixg=FALSE, the g specifies the initial value of g used before optimization.}

\item{fixg}{If TRUE, don't estimate g but use the specified g.}

\item{output}{List of values to be returned. Defaults to "result" (summary results), "fitted_g"
(the fitted prior), and loglik.}
}
\value{
a list with elements specified by output parameter
}
\description{
Paragraph-length description goes here.
}
\details{
Given vectors of data x, and standard errors s, solve EBNM problem with "point-laplace" prior
(i.e. a mixture of point mass at 0 and Laplace distribution).
That is the prior is \eqn{pi0 \delta_0 + (1-pi0)DExp(a)} where Dexp is the double exponential
(Laplace) distribution, and (pi0,a) are esimated by marginal maximum likelihood. The model and code are based on
EbayesThresh by Johnstone and Silverman, but simplified by removing thresholding, and
improved numerical stability.
}
\examples{
mu = c(rep(0,1000), rexp(1000)) # means
s = rgamma(2000,1,1) #standard errors
x = mu + rnorm(2000,0,s) # observations
x.ebnm = ebnm_point_laplace(x,s)
ashr::get_pm(x.ebnm) # posterior mean

}
