% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_point_normal.R
\name{ebnm_point_normal}
\alias{ebnm_point_normal}
\title{Solve the Empirical Bayes Normal Means problem with point-normal prior}
\usage{
ebnm_point_normal(x, s = 1, g = NULL, fixg = FALSE, norm = mean(s))
}
\arguments{
\item{x}{a vector of observations}

\item{s}{a vector of standard deviations (or scalar if all equal)}

\item{g}{The prior distribution (list with elements pi0,a). Usually this is unspecified (NULL) and
estimated from the data. However, it can be used in conjuction with fixg=TRUE
to specify the g to use (e.g. useful in simulations to do computations with the "true" g).
Or, if g is specified but fixg=FALSE, the g specifies the initial value of g used before optimization.}

\item{fixg}{If TRUE, don't estimate g but use the specified g.}

\item{norm}{normalization factor to divide x and s by before running optimization (should not affect results, but improves numerical stability when x and s are tiny)}
}
\value{
a list with elements result, fitted_g, and loglik
}
\description{
Paragraph-length description goes here.
}
\details{
Given vectors of data x, and standard errors s, solve EBNM problem with "point-normal" prior
(i.e. a mixture of point mass at 0 and normal distribution).
That is the prior is \eqn{pi0 \delta_0 + (1-pi0)N(0,1/a)} where N is the normal
distribution, and (pi0,a) are estimated by marginal maximum likelihood.
}
\examples{
mu = c(rep(0,1000), rexp(1000)) # means
s = rgamma(2000,1,1) #standard errors
x = mu + rnorm(2000,0,s) # observations
x.ebnm = ebnm_point_normal(x,s)
ashr::get_pm(x.ebnm) # posterior mean

}
