% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_exponential.R
\name{exponentialmix}
\alias{exponentialmix}
\title{Constructor for exponentialmix class}
\usage{
exponentialmix(pi, mode, scale)
}
\arguments{
\item{pi}{A vector of mixture proportions.}

\item{scale}{A vector of scale parameters.}

\item{mean}{A vector of modes.}
}
\description{
Creates a finite mixture of exponential distributions.
}
