% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{fitted.ebnm}
\alias{fitted.ebnm}
\title{Extract the posterior means from an ebnm object}
\usage{
\method{fitted}{ebnm}(object, ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\value{
The posterior means \eqn{\hat{\theta}} in a fitted \code{ebnm} model.
}
\description{
The \code{\link[stats]{fitted}} method for class \code{ebnm}.
}
