% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{fitted.ebnm}
\alias{fitted.ebnm}
\title{Extract posterior estimates from a fitted EBNM model}
\usage{
\method{fitted}{ebnm}(object, ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\description{
The \code{\link[stats]{fitted}} method for class \code{\link{ebnm}}.
  Returns a data frame that includes posterior means, standard deviations,
  and local false sign rates (when available).
}
