% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_exponential.R
\name{gammamix}
\alias{gammamix}
\title{Constructor for gammamix class}
\usage{
gammamix(pi, shape, scale, shift = rep(0, length(pi)))
}
\arguments{
\item{pi}{A vector of mixture proportions.}

\item{shape}{A vector of shape parameters.}

\item{scale}{A vector of scale parameters.}

\item{shift}{A vector of shift parameters.}
}
\value{
An object of class \code{gammamix} (a list with elements
  \code{pi}, \code{shape}, \code{scale}, and \code{shift}, described above).
}
\description{
Creates a finite mixture of gamma distributions.
}
