% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_laplace.R
\name{laplacemix}
\alias{laplacemix}
\title{Constructor for laplacemix class}
\usage{
laplacemix(pi, mean, scale)
}
\arguments{
\item{pi}{A vector of mixture proportions.}

\item{mean}{A vector of means.}

\item{scale}{A vector of scale parameters.}
}
\value{
Describe the return value here.
}
\description{
Creates a finite mixture of Laplace distributions.
}
