% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/point_laplace.R
\name{laplacemix}
\alias{laplacemix}
\title{Constructor for laplacemix class}
\usage{
laplacemix(pi, mean, scale)
}
\arguments{
\item{pi}{A vector of mixture proportions.}

\item{mean}{A vector of means.}

\item{scale}{A vector of scale parameters.}
}
\value{
An object of class \code{laplacemix} (a list with elements
  \code{pi}, \code{mean}, and \code{scale}, described above).
}
\description{
Creates a finite mixture of Laplace distributions.
}
