% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{print.ebnm}
\alias{print.ebnm}
\title{Print an ebnm object}
\usage{
\method{print}{ebnm}(x, ...)
}
\arguments{
\item{x}{The fitted \code{ebnm} object.}

\item{...}{Not used. Included for consistency as an S3 method.}
}
\description{
The \code{print} method for class \code{ebnm}.
}
