% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebnm_methods.R
\name{samp.ebnm}
\alias{samp.ebnm}
\title{Sample from the posterior of a fitted EBNM model}
\usage{
\method{samp}{ebnm}(object, nsamp, ...)
}
\arguments{
\item{object}{The fitted \code{ebnm} object.}

\item{nsamp}{The number of posterior samples to return per observation.}

\item{...}{Additional arguments to be passed to the posterior sampler
function. Since \code{ebnm_horseshoe} returns an MCMC sampler, it takes
parameter \code{burn}, the number of burn-in samples to discard.  At
present, no other samplers take any additional parameters.}
}
\value{
A matrix of posterior samples, with rows corresponding to
  distinct samples and columns corresponding to observations.
}
\description{
A convenience function that extracts the posterior sampler from an object
  of class \code{\link{ebnm}} and returns a specified number of samples.
}
